#58 Update build files - Use javafxplugin - Moved all plugins to root build - Updated license plugin - Moved samples classes to avoid split packages
