+++ FXyz-Client/build.gradle
@@ -1,29 +1,11 @@
-buildscript {
-    repositories {
-        jcenter()
-    }
-    dependencies {
-        classpath 'com.github.jengelman.gradle.plugins:shadow:1.2.4'
-    }
-}
-
-plugins {
-    //id 'com.github.hierynomus.license' version '0.13.1'
-    id 'maven-publish'
-    id 'net.nemerosa.versioning' version '2.6.1'
-    id 'com.jfrog.bintray' version '1.8.4'
-    id 'com.github.ben-manes.versions' version '0.13.0'
-}
-
+apply plugin: 'maven-publish'
 apply from: 'gradle/publishing.gradle'
 
 repositories {
     jcenter()
 }
 
-import com.github.jengelman.gradle.plugins.shadow.transformers.ServiceFileTransformer
 version = fxyz_sampler_version
-apply plugin: 'com.github.johnrengelman.shadow'
 
 sourceCompatibility = '1.9'
 targetCompatibility = '1.9'
@@ -39,16 +21,6 @@ dependencies {
         compile 'org.controlsfx:controlsfx:8.40.14'
     }
     compile 'org.reactfx:reactfx:2.0-M5'    
-}
-
-compileJava {
-    doFirst {
-        options.compilerArgs = [
-                '--module-path', classpath.asPath,
-                '--add-modules', 'javafx.controls,javafx.fxml,javafx.swing'
-        ]
-    }
-    options.encoding = "UTF-8"
 }
 
 javadoc {
@@ -79,9 +51,6 @@ shadowJar {
     mergeServiceFiles {
         include 'META-INF/services/org.fxyz3d.*'
     }
-    //transform(ServiceFileTransformer) {
-    //path = 'META-INF/services/org.fxyz3d.*'
-    //}
 }
 // sign all generated jars
 //signing {
