+++ FXyz-Core/src/main/java/org/fxyz3d/shapes/polygon/SubdivisionMesh.java
@@ -1,4 +1,3 @@
-
 /*
  * Copyright (c) 2010, 2014, Oracle and/or its affiliates.
  * All rights reserved. Use is subject to license terms.
