+++ FXyz-Samples/src/main/java/org/fxyz3d/samples/HomePage.java
@@ -27,7 +27,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */ 
 
-package org.fxyz3d;
+package org.fxyz3d.samples;
 
 import javafx.scene.Node;
 import org.fxyz3d.model.WelcomePage;
