+++ FXyz-Samples/src/main/java/org/fxyz3d/samples/Launcher.java
@@ -26,7 +26,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */ 
-package org.fxyz3d;
+package org.fxyz3d.samples;
 
 import org.fxyz3d.client.FXyzClient;
