+++ FXyz-Samples/src/main/resources/META-INF/services/org.fxyz3d.FXyzSamplerProject
@@ -1,1 +1,1 @@
-org.fxyz3d.FXyzProject
+org.fxyz3d.samples.FXyzProject
